requirements.txt
Django==2.2
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.7.0
Markdown==3.2.2
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
Django==2.2
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==1.7.0
Markdown==3.2.2
mysqlclient==2.0.1
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
